requirements.txt
pysha3>=1.02
